refactor(ui): synced only is default and uses local storage
