+++ services/dashboard-ui/client/components/common/SyncedFilter/SyncedFilterContainer.tsx
@@ -1,21 +1,12 @@
 import { type ChangeEvent } from 'react'
-import { useNavigate, useSearchParams } from 'react-router'
+import { useSyncedOnlyFilter } from '@/hooks/use-synced-only-filter'
 import { SyncedFilter } from './SyncedFilter'
 
 export const SyncedFilterContainer = () => {
-  const navigate = useNavigate()
-  const [searchParams] = useSearchParams()
-  const syncedOnly = searchParams.get('synced_only') === 'true'
+  const { syncedOnly, setSyncedOnly } = useSyncedOnlyFilter()
 
   const handleChange = (e: ChangeEvent<HTMLInputElement>) => {
-    const params = new URLSearchParams(searchParams.toString())
-    if (e.target.checked) {
-      params.set('synced_only', 'true')
-    } else {
-      params.delete('synced_only')
-    }
-    params.delete('offset')
-    navigate(`?${params.toString()}`, { replace: true })
+    setSyncedOnly(e.target.checked)
   }
 
   return <SyncedFilter syncedOnly={syncedOnly} onChange={handleChange} />
